# TaskManager-Api-Swagger.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Manager API (Flask)",
    "version": "1.0.0",
    "description": "A RESTful API with JWT auth, roles (admin/user), CRUD on tasks, pagination and filtering."
  },
  "servers": [
    { "url": "http://127.0.0.1:5000", "description": "Local" }
  ],
  "tags": [
    { "name": "Auth", "description": "User registration and login" },
    { "name": "Tasks", "description": "Task CRUD, pagination & filtering" }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "operationId": "registerUser",
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "examples": {
                "user": {
                  "value": {
                    "email": "user@example.com",
                    "password": "secret123"
                  }
                },
                "admin": {
                  "value": {
                    "email": "admin@example.com",
                    "password": "secret123",
                    "role": "admin"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input / email already registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "loginUser",
        "summary": "Login and get JWT access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "examples": {
                "default": {
                  "value": {
                    "email": "user@example.com",
                    "password": "secret123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "operationId": "listTasks",
        "summary": "List tasks (pagination & filtering)",
        "description": "Authentication optional. If unauthenticated, returns an empty list by design.",
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/PerPageParam" },
          {
            "name": "completed",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean" },
            "description": "Filter by completion status"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of tasks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedTasks" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "operationId": "createTask",
        "summary": "Create a new task",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskCreate" },
              "examples": {
                "default": {
                  "value": {
                    "title": "Buy milk",
                    "description": "2L whole milk",
                    "completed": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "operationId": "getTask",
        "summary": "Get a task by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TaskIdPath" }],
        "responses": {
          "200": {
            "description": "Task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "operationId": "updateTask",
        "summary": "Update a task by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TaskIdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskUpdate" },
              "examples": {
                "default": {
                  "value": {
                    "title": "New title",
                    "completed": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "operationId": "deleteTask",
        "summary": "Delete a task by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TaskIdPath" }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supply `Authorization: Bearer <token>`"
      }
    },
    "parameters": {
      "TaskIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 },
        "description": "Task ID"
      },
      "PageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "default": 1 },
        "description": "Page number (default 1)"
      },
      "PerPageParam": {
        "name": "per_page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 },
        "description": "Items per page (default 10, max 100)"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Not authorized to access this resource",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "minLength": 6, "example": "secret123" },
          "role": { "type": "string", "enum": ["user", "admin"], "default": "user" }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["user", "admin"] }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "secret123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string", "description": "JWT Bearer token" }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 42 },
          "title": { "type": "string", "example": "Buy milk" },
          "description": { "type": "string", "example": "2L whole milk" },
          "completed": { "type": "boolean", "example": false },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "TaskCreate": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string", "default": "" },
          "completed": { "type": "boolean", "default": false }
        }
      },
      "TaskUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string" },
          "completed": { "type": "boolean" }
        }
      },
      "PaginatedTasks": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Task" }
          },
          "page": { "type": "integer", "example": 1 },
          "per_page": { "type": "integer", "example": 10 },
          "total": { "type": "integer", "example": 25 }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Invalid credentials" },
          "errors": { "type": "object", "additionalProperties": true }
        }
      }
    }
  }
}